+++ Write-FormatViewExpression.ps1
@@ -87,10 +87,31 @@
     [switch]
     $Underline,
 
+    # If set, will double underline the -Text, -Property, or -ScriptBlock.
+    # This is only valid in consoles that support ANSI terminals, or in HTML
+    [switch]
+    $DoubleUnderline,
+
+    # If set, make the -Text, -Property, or -ScriptBlock Italic.
+    # This is only valid in consoles that support ANSI terminals, or in HTML
+    [Alias('Italics')]
+    [switch]
+    $Italic,
+
+    # If set, will hide  the -Text, -Property, or -ScriptBlock.
+    # This is only valid in consoles that support ANSI terminals, or in HTML
+    [switch]
+    $Hide,
+
     # If set, will invert the -Text, -Property, -or -ScriptBlock
     # This is only valid in consoles that support ANSI terminals, or in HTML.
     [switch]
     $Invert,
+
+    # If set, will cross out the -Text, -Property, -or -ScriptBlock
+    # This is only valid in consoles that support ANSI terminals, or in HTML.
+    [Alias('Strikethrough', 'Crossout')]
+    [switch]$Strikethru,
 
     # If provided, will output the format using this format string.
     [string]
@@ -145,7 +166,16 @@
         }
 
         foreach ($n in 1..$count) {
-            if ($ForegroundColor -or $BackgroundColor -or $Bold -or $Underline) {                
+            if ($ForegroundColor -or 
+                $BackgroundColor -or 
+                $Bold -or 
+                $Underline -or 
+                $Italic -or 
+                $Faint -or 
+                $doubleUnderline -or 
+                $Invert -or
+                $Strikethru -or
+                $hide) {                
                 $colorize = [ScriptBlock]::Create("@(Format-RichText $(@(
                     if ($ForegroundColor) {
                         "-ForegroundColor '$ForeGroundColor'"
@@ -153,9 +183,14 @@
                     if ($BackgroundColor) {
                         "-BackgroundColor '$BackgroundColor'"
                     }
+                    if ($Italic) { '-Italic' }
                     if ($Bold) { '-Bold' }
+                    if ($Faint) { '-Faint' }
                     if ($Underline) { '-Underline'}
+                    if ($DoubleUnderline) { '-DoubleUnderline'}
+                    if ($hide) { '-Hide' }
                     if ($Invert) { '-Invert' }
+                    if ($Strikethru) { '-Strikethru' }
                     '-NoClear'
                 ) -join ' ')) -join ''")
                 Write-FormatViewExpression -ScriptBlock $colorize
@@ -201,8 +236,17 @@ $if")
                 "$xOut".Substring('<?xml version="1.0" encoding="utf-16"?>'.Length + [Environment]::NewLine.Length)
                 $xOut.Dispose()
             }
-            if ($ForegroundColor -or $BackgroundColor -or $Bold -or $Underline -or $Invert) {
-
+            if ($ForegroundColor -or 
+                $BackgroundColor -or 
+                $Bold -or 
+                $Underline -or 
+                $Italic -or 
+                $Faint -or 
+                $doubleUnderline -or 
+                $Invert -or
+                $Hide -or 
+                $Strikethru
+            ) {
                 Write-FormatViewExpression -ScriptBlock ([ScriptBlock]::Create(($colorize -replace '-NoClear')))
             }
         }